Migrate Diagnostics test to new expectations runner
